Rename files according to naming convention
