+++ client/src/css/availability-page.css
