+++ client/src/css/availability-picker.css
